pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minion-factory"
version = "0.3.45"
description = "Unified multi-agent coordination framework — CLI, daemon, task DAG"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
]

[project.scripts]
minion = "minion.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/minion"]

[tool.hatch.build.targets.wheel.force-include]
"docs" = "minion/docs"
"missions" = "minion/missions"
"task-flows" = "minion/task-flows"
"src/minion/prompts/capabilities" = "minion/prompts/capabilities"
"src/minion/prompts/roles" = "minion/prompts/roles"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
